add order

+++ src/main/java/cn/hlvan/constant/UserType.java
@@ -5,20 +5,20 @@ public class UserType {
     /**
      * 超级管理员
      */
-    public static Integer ADMIN = 1;
+    public static byte ADMIN = 1;
 
     /**
      * 部门管理员
      */
-    public static Integer MANAGER = 2;
+    public static byte MANAGER = 2;
 
     /**
      * 商家
      */
-    public static Integer MERCHANT = 3;
+    public static byte MERCHANT = 3;
     /**
      * 商家
      */
-    public static Integer WRITER = 4;
+    public static byte WRITER = 4;
 
 }
